Added a little more functionality, made a dev build of the game complete with reset functionality!
