+++ Assets/Scripts/CameraShake.cs.meta
