+++ Assets/Scripts/PlayerScripts/PlayerController.cs
@@ -192,7 +192,7 @@ public class PlayerController : MonoBehaviour {
     }
     void Triangle_Button ()
     {
-        SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
+        //SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
     }
     void Square_Button ()
     {
